testing: Warm up rustdoc cache ahead of test execution, thus minimizing contention

+++ ui-tests/tests/cbindgen/rust/cases/constant/src/lib.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-pub const FOO: i32 = 10;
+pub const FOO: usize = 10;
 pub const BAR: &'static str = "hello world";
 pub const DELIMITER: char = ':';
 pub const LEFTCURLY: char = '{';
